moving logic to the useMarkdownInput hook
